requirements.txt
torch~=1.13.1+cu117
torch-geometric
numpy~=1.24.4
pandas~=2.2.3
nltk~=3.9.1
scikit-learn~=1.5.2
pyrootutils~=1.0.4
tqdm~=4.66.5
swifter
omegaconf~=2.3.0
wandb~=0.18.1
hydra-core
transformers~=4.24.0
scipy~=1.8.0
spacy~=3.8.3
matplotlib~=3.9.4
seaborn~=0.13.2
NewsSentiment~=1.2.28